requirements.txt
Django == 4.1.1
djangorestframework
PyJWT